src/app/reactiveForm/reactiveForm.component.spec.ts
import { TestBed, async } from '@angular/core/testing';
import { ReactiveForm } from './reactiveForm.component';

describe('ReactiveForm', () => {
  beforeEach(() => {
    TestBed.configureTestingModule({
      declarations: [
        ReactiveForm
      ],
    });
    TestBed.compileComponents();
  });

  it('should create the app', async(() => {
    const fixture = TestBed.createComponent(ReactiveForm);
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  }));

  it(`should have as title 'app works!'`, async(() => {
    const fixture = TestBed.createComponent(ReactiveForm);
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('app works!');
  }));

  it('should render title in a h1 tag', async(() => {
    const fixture = TestBed.createComponent(ReactiveForm);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('app works!');
  }));
});
